Move chat sending logic to input component
